Currently experiencing a cacheing issue.  When the program is being run, the data is appearing as if its already there.

+++ src/main/java/org/perscholas/services/EventService.java
@@ -29,7 +29,7 @@ public class EventService {
     }
 
 
-    public void updateEvent(String eventType, Long eventEmployeeNumber, Double eventCost, Double eventDuration){
+    public void updateEvent(String eventType, Double eventEmployeeNumber, Double eventCost, Double eventDuration){
 
         boolean exists = eventRepo.findEventByeventType(eventType).isPresent();//.existsById(eventType);
         if(exists){
